% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Hello}
\alias{Hello}
\title{Hello}
\usage{
Hello(MyMessage = "Hello")
}
\arguments{
\item{MyMessage}{String corresponding to the message to display}
}
\value{
a Message
}
\description{
Hello
}
\examples{
Hello()
Hello(MyMessage = 'Bonjour')

}
