% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PracticeThinkR-package.R
\docType{package}
\name{PracticeThinkR-package}
\alias{PracticeThinkR}
\alias{PracticeThinkR-package}
\title{PracticeThinkR: Practice ThinkR Training}
\description{
Blabla.
}
\author{
\strong{Maintainer}: Aurélien Macé \email{aurelien.mace@finddx.org}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
